Apply fusion optimization to input iterator
