+++ tests/beman/any_view/constexpr.test.cpp
@@ -65,10 +65,20 @@ TEST(ConstexprTest, sort_vector) {
     EXPECT_TRUE(sort(std::vector{6, 8, 7, 5, 3, 0, 9}));
 }
 
+template <class T>
+struct non_trivially_copyable {
+    T value;
+
+    non_trivially_copyable(const non_trivially_copyable&);
+};
+
+template <class T>
+non_trivially_copyable<T>::non_trivially_copyable(const non_trivially_copyable&) = default;
+
 constexpr auto set_front(any_view<int, forward> view, int value) {
-    // forward iterator of lvalue reference uses cache object to fuse virtual dispatches
-    static_assert(sizeof(std::ranges::iterator_t<any_view<int, forward>>) ==
-                  sizeof(std::ranges::iterator_t<any_view<int, input>>) + sizeof(int*));
+    // lvalue reference uses cache object to fuse virtual dispatches
+    static_assert(sizeof(std::ranges::iterator_t<any_view<int>>) ==
+                  sizeof(std::ranges::iterator_t<proxy_any_view<non_trivially_copyable<int>>>) + sizeof(int*));
 
     auto& ref = view.front();
     // even with cache object, lifetime of reference is not tied to lifetime of iterator
